fix: Async가 무시되고 동기적으로 실행되는 문제 수정 - 분리했더니 괜찮아졌지만, 이유는 모르겠음
